refactor(tx3-cardano): extract script data hash into its own module
Move `script_ref_language_key`, `collect_used_plutus_versions` and
`compute_script_data_hash` -- plus the tests that exercise them -- out of
the large `compile/mod.rs` into a focused `compile/script_data.rs`. The
three functions form one cohesive responsibility (the script integrity
hash) with a single caller (`entry_point`), so the seam is clean.

`compute_script_data_hash` is now `pub(super)` and the helpers are private
to the module; the tests are self-contained (a trivial empty-cost-model
`PParams`), so they no longer depend on the integration test harness.
