+++ crates/tx3-cardano/src/tests.rs
@@ -548,290 +548,3 @@ async fn min_utxo_compiler_op_test() {
         assert!(!assets.is_empty());
     }
 }
-
-mod script_language_collection {
-    use super::*;
-    use std::collections::BTreeSet;
-
-    use pallas::codec::{
-        minicbor,
-        utils::{CborWrap, KeepRaw, NonEmptySet},
-    };
-    use pallas::ledger::primitives::conway::{NativeScript, ScriptRef};
-    use pallas::ledger::primitives::PlutusScript;
-    use proptest::prelude::*;
-
-    fn empty_tx() -> tir::Tx {
-        tir::Tx {
-            fees: tir::Expression::None,
-            references: vec![],
-            inputs: vec![],
-            outputs: vec![],
-            validity: None,
-            mints: vec![],
-            burns: vec![],
-            adhoc: vec![],
-            collateral: vec![],
-            signers: None,
-            metadata: vec![],
-        }
-    }
-
-    fn empty_witness_set() -> primitives::conway::WitnessSet<'static> {
-        primitives::conway::WitnessSet {
-            redeemer: None,
-            vkeywitness: None,
-            native_script: None,
-            bootstrap_witness: None,
-            plutus_data: None,
-            plutus_v1_script: None,
-            plutus_v2_script: None,
-            plutus_v3_script: None,
-        }
-    }
-
-    fn utxo_with_script(script: Option<tir::Expression>) -> UtxoSet {
-        let utxo = Utxo {
-            r#ref: UtxoRef {
-                txid: vec![0u8; 32],
-                index: 0,
-            },
-            address: vec![],
-            assets: CanonicalAssets::from_naked_amount(1),
-            datum: None,
-            script,
-        };
-
-        [utxo].into()
-    }
-
-    fn ref_script_expr(script_ref: &ScriptRef) -> tir::Expression {
-        tir::Expression::Bytes(minicbor::to_vec(script_ref).unwrap())
-    }
-
-    fn plutus_v2() -> ScriptRef<'static> {
-        ScriptRef::PlutusV2Script(PlutusScript::<2>(vec![0x01, 0x02, 0x03].into()))
-    }
-
-    fn plutus_v3() -> ScriptRef<'static> {
-        ScriptRef::PlutusV3Script(PlutusScript::<3>(vec![0x01, 0x02, 0x03].into()))
-    }
-
-    fn native() -> ScriptRef<'static> {
-        let script = NativeScript::ScriptPubkey(primitives::Hash::<28>::from([7u8; 28]));
-        ScriptRef::NativeScript(KeepRaw::from(script))
-    }
-
-    // A reference input carrying a PlutusV2 reference script, with no inline
-    // witness scripts, must yield the V2 language key {1} -- explicitly NOT the
-    // old PlutusV3 default {2}.
-    #[test]
-    fn reference_only_plutus_v2() {
-        let mut tx = empty_tx();
-        tx.references = vec![tir::Expression::UtxoSet(utxo_with_script(Some(
-            ref_script_expr(&plutus_v2()),
-        )))];
-
-        let versions =
-            crate::compile::collect_used_plutus_versions(&tx, &empty_witness_set()).unwrap();
-
-        assert_eq!(versions, BTreeSet::from([1]));
-        assert!(!versions.contains(&2));
-    }
-
-    // A native reference script contributes no language view.
-    #[test]
-    fn native_reference_script_is_ignored() {
-        let mut tx = empty_tx();
-        tx.references = vec![tir::Expression::UtxoSet(utxo_with_script(Some(
-            ref_script_expr(&native()),
-        )))];
-
-        let versions =
-            crate::compile::collect_used_plutus_versions(&tx, &empty_witness_set()).unwrap();
-
-        assert!(versions.is_empty());
-
-        // With no plutus scripts and no redeemers there is no script data hash.
-        let hash = crate::compile::compute_script_data_hash(
-            &tx,
-            &empty_witness_set(),
-            &test_compiler(None).pparams,
-        )
-        .unwrap();
-        assert!(hash.is_none());
-    }
-
-    // Property: the collected language set equals exactly the set of Plutus
-    // language keys of all scripts present -- across witness scripts, spending
-    // inputs and reference inputs -- with native scripts excluded and duplicates
-    // deduped, regardless of how the scripts are distributed across sources.
-    #[derive(Clone, Copy, Debug)]
-    enum Lang {
-        Native,
-        V1,
-        V2,
-        V3,
-    }
-
-    impl Lang {
-        fn script_ref(self, seed: u8) -> ScriptRef<'static> {
-            match self {
-                Lang::Native => ScriptRef::NativeScript(KeepRaw::from(NativeScript::ScriptPubkey(
-                    primitives::Hash::<28>::from([seed; 28]),
-                ))),
-                Lang::V1 => ScriptRef::PlutusV1Script(PlutusScript::<1>(vec![seed].into())),
-                Lang::V2 => ScriptRef::PlutusV2Script(PlutusScript::<2>(vec![seed].into())),
-                Lang::V3 => ScriptRef::PlutusV3Script(PlutusScript::<3>(vec![seed].into())),
-            }
-        }
-
-        /// The expected `LanguageViews` key per the spec (native has none).
-        fn key(self) -> Option<PlutusVersion> {
-            match self {
-                Lang::Native => None,
-                Lang::V1 => Some(0),
-                Lang::V2 => Some(1),
-                Lang::V3 => Some(2),
-            }
-        }
-    }
-
-    #[derive(Clone, Copy, Debug)]
-    enum Source {
-        Witness,
-        SpendingInput,
-        ReferenceInput,
-    }
-
-    fn utxo_set_with_ref_script(seed: u8, script_ref: &ScriptRef) -> UtxoSet {
-        let utxo = Utxo {
-            // Unique ref per seed so the UtxoSet does not dedup distinct scripts.
-            r#ref: UtxoRef {
-                txid: vec![seed; 32],
-                index: 0,
-            },
-            address: vec![],
-            assets: CanonicalAssets::from_naked_amount(1),
-            datum: None,
-            script: Some(ref_script_expr(script_ref)),
-        };
-
-        [utxo].into()
-    }
-
-    fn build_tx(
-        placements: &[(Source, Lang)],
-    ) -> (tir::Tx, primitives::conway::WitnessSet<'static>) {
-        let mut tx = empty_tx();
-        let mut witness_set = empty_witness_set();
-
-        let mut native_ws = vec![];
-        let mut v1_ws = vec![];
-        let mut v2_ws = vec![];
-        let mut v3_ws = vec![];
-
-        for (i, (source, lang)) in placements.iter().enumerate() {
-            let seed = i as u8;
-            match source {
-                Source::Witness => match lang {
-                    Lang::Native => native_ws.push(KeepRaw::from(NativeScript::ScriptPubkey(
-                        primitives::Hash::<28>::from([seed; 28]),
-                    ))),
-                    Lang::V1 => v1_ws.push(PlutusScript::<1>(vec![seed].into())),
-                    Lang::V2 => v2_ws.push(PlutusScript::<2>(vec![seed].into())),
-                    Lang::V3 => v3_ws.push(PlutusScript::<3>(vec![seed].into())),
-                },
-                Source::SpendingInput => tx.inputs.push(tir::Input {
-                    name: format!("input_{i}"),
-                    utxos: tir::Expression::UtxoSet(utxo_set_with_ref_script(
-                        seed,
-                        &lang.script_ref(seed),
-                    )),
-                    redeemer: tir::Expression::None,
-                }),
-                Source::ReferenceInput => {
-                    tx.references
-                        .push(tir::Expression::UtxoSet(utxo_set_with_ref_script(
-                            seed,
-                            &lang.script_ref(seed),
-                        )))
-                }
-            }
-        }
-
-        witness_set.native_script = NonEmptySet::from_vec(native_ws);
-        witness_set.plutus_v1_script = NonEmptySet::from_vec(v1_ws);
-        witness_set.plutus_v2_script = NonEmptySet::from_vec(v2_ws);
-        witness_set.plutus_v3_script = NonEmptySet::from_vec(v3_ws);
-
-        (tx, witness_set)
-    }
-
-    fn lang_strategy() -> impl Strategy<Value = Lang> {
-        prop_oneof![
-            Just(Lang::Native),
-            Just(Lang::V1),
-            Just(Lang::V2),
-            Just(Lang::V3),
-        ]
-    }
-
-    fn source_strategy() -> impl Strategy<Value = Source> {
-        prop_oneof![
-            Just(Source::Witness),
-            Just(Source::SpendingInput),
-            Just(Source::ReferenceInput),
-        ]
-    }
-
-    proptest! {
-        #[test]
-        fn collected_versions_match_present_languages(
-            placements in prop::collection::vec((source_strategy(), lang_strategy()), 0..8)
-        ) {
-            let (tx, witness_set) = build_tx(&placements);
-
-            let expected: BTreeSet<PlutusVersion> =
-                placements.iter().filter_map(|(_, lang)| lang.key()).collect();
-
-            let got = crate::compile::collect_used_plutus_versions(&tx, &witness_set).unwrap();
-
-            prop_assert_eq!(got, expected);
-        }
-    }
-
-    // A used language with no configured cost model surfaces an error rather
-    // than panicking (the old `.unwrap()`).
-    #[test]
-    fn missing_cost_model_errors() {
-        let mut tx = empty_tx();
-        tx.references = vec![tir::Expression::UtxoSet(utxo_with_script(Some(
-            ref_script_expr(&plutus_v3()),
-        )))];
-
-        // Redeemers must be present for build_for to attach language views, but
-        // the missing cost model is detected before that. Use pparams with an
-        // empty cost model map so key 2 is missing.
-        let mut pparams = test_compiler(None).pparams;
-        pparams.cost_models.clear();
-
-        let result = crate::compile::compute_script_data_hash(&tx, &empty_witness_set(), &pparams);
-        assert!(matches!(
-            result,
-            Err(tx3_tir::compile::Error::FormatError(_))
-        ));
-    }
-
-    // A reference script double-wrapped in a CborWrap still decodes.
-    #[test]
-    fn cbor_wrap_fallback() {
-        let wrapped = tir::Expression::Bytes(minicbor::to_vec(CborWrap(plutus_v2())).unwrap());
-
-        let script_ref = crate::coercion::expr_into_script_ref(&wrapped).unwrap();
-        assert_eq!(
-            crate::compile::script_ref_language_key(&script_ref),
-            Some(1)
-        );
-    }
-}
